Add PowerFunction method to calculate exponentiation
Implemented a PowerFunction(int x, int n) method that calculates x raised to the power n using an iterative loop.

The method multiplies x with itself n times and prints the result in a formatted message.

Added a main method that takes two integer inputs (x and n) from the user using Scanner and passes them to the PowerFunction.

Example: If input is 2 5, output will be 2 raised to the power 5 is :32.